+++ packages/graphql-yoga/CHANGELOG.md
@@ -1,5 +1,17 @@
 # graphql-yoga
 
+## 3.0.0-next.4
+
+### Major Changes
+
+- [#1808](https://github.com/dotansimha/graphql-yoga/pull/1808) [`02d2aecd`](https://github.com/dotansimha/graphql-yoga/commit/02d2aecdee55e4c54454c48c2ca0fd7425796ae0) Thanks [@enisdenjo](https://github.com/enisdenjo)! - Drop `readinessCheckEndpoint` and introduce `useReadinessCheck` plugin
+
+### Patch Changes
+
+- [#1844](https://github.com/dotansimha/graphql-yoga/pull/1844) [`b079c93b`](https://github.com/dotansimha/graphql-yoga/commit/b079c93ba47dc94d58f7d2b738a9423c29a149a1) Thanks [@ardatan](https://github.com/ardatan)! - - All unexpected errors even if they are masked/wrapped, HTTP status code will be set to 500.
+
+  > "Unexpected error" means an Error that is not an instance of GraphQLError or an instance of GraphQLError with an `originalError` that is not an instance of GraphQLError recursively.
+
 ## 3.0.0-next.3
 
 ### Patch Changes